requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
argcomplete==2.0.0
attrs==23.2.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blinker==1.7.0
blivet==3.8.2
blivet-gui==2.5.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dataclasses-json==0.6.4
dbus-python==1.3.2
discord.py==2.4.0
distro==1.8.0
dnf==4.21.0
docopt==0.6.2
fb-re2==1.0.7
fedora-third-party==0.10
feedparser==6.0.11
file-magic==0.4.0
Flask==3.0.3
Flask-Cors==4.0.0
fros==1.1
frozenlist==1.4.1
gpg==1.20.0
greenlet==3.0.3
h11==0.14.0
humanize==3.13.1
idna==3.7
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-text-splitters==0.0.1
langsmith==0.1.48
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.2
lxml==4.9.3
Mako==1.2.3
Markdown==3.5.2
MarkupSafe==2.1.3
marshmallow==3.21.1
mercurial==6.5.3
mistune==3.0.2
multidict==6.0.5
mutagen==1.46.0
mypy-extensions==1.0.0
nftables==0.1
nicotine-plus==3.3.4
numpy==1.26.4
olefile==0.46
orjson==3.10.1
packaging==23.2
Paste==3.5.3
pdf2image==1.17.0
pexpect==4.8.0
pid==2.2.3
pillow==10.3.0
ply==3.11
productmd==1.38
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.19.0
pycryptodomex==3.20.0
pycups==2.0.4
pydantic==2.7.0
pydantic_core==2.18.1
pyenchant==3.2.2
PyGObject==3.46.0
pykickstart==3.48
pyOpenSSL==23.2.0
pyparted==3.13.0
PyQt5==5.15.9
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-vlc==3.0.20123
pythondialog==3.5.3
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2024.4.16
reportlab==4.1.0
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.19.1.1
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
sgmllib3k==1.0.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.6.0
soupsieve==2.5
SQLAlchemy==2.0.29
systemd-python==235
Tempita==0.5.2
tenacity==8.2.3
torbrowser-launcher==0.3.7
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==1.26.19
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==3.0.2
yarl==1.9.4
yt-dlp==2024.8.6
zapzap==5.3
zombie-imp==0.0.2
